+++ com.alibaba.testme.service/src/main/java/com/alibaba/testme/service/WorkSpaceService.java
@@ -15,7 +15,7 @@ public interface WorkSpaceService {
      * @param workSpaceDO
      * @return
      */
-    public Long addWorkSpaceDO(WorkSpaceDO workSpaceDO);
+    public Integer addWorkSpaceDO(WorkSpaceDO workSpaceDO);
 
     /**
      * @param DO
